cerulean_cloud/cloud_function_ais_analysis/requirements.txt
asyncio==3.4.3
json==Package(s) not found
os==Package(s) not found
pandas_gbq==0.17.6
google-cloud-tasks==2.9.1
google-cloud-core==2.3.1
google-auth==1.35.0
sqlalchemy==1.4.32
datetime==Package(s) not found
math==Package(s) not found
centerline==1.0.1
geopandas==0.12.2
movingpandas==0.15rc1
numpy==1.21.5
scipy==1.8.0
shapely==2.0.1
earthengine-api==0.1.342